Add files via upload
USB-Soc-Eng-Demo.bat can be placed on a Thumb Drive in a folder obscured however you wish or in this example just "Files"
We then entice the user to click for example "Salary Reviews 2017.ink" that is using a PDF icon or similar 

When opened the USB will launch IE , requesting a site of your choosing and a unique ID to id the USB or user?
This is how we can tell the USB was plugged in and files accessed without fancy hardware.
Whilst this is happening we demonstrate to the User/Client how we can then Create files and folders on the users Home Directory and even create a Zip of those files. This stands out because to demonstrate the danger we do not always need a reverse shell. 
Much Ransom-ware does not dial home at all, It just silently starts encrypting files on the users drive under the users permissions..

Thoughts?